Code cleaning pt 2: controls, fire, sprint
I got rid of the dummy copy/paste I did from the internet.

Player can walk or run. It depends on a base value that defines the walk and run speed by a ratio.

Firing code is back too with a cooldown between each shot. Next step is to trigger the Factory from go.get_rotation of the player.